requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
decorator==5.2.1
distro==1.9.0
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
google-auth==2.40.3
google-genai==1.28.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.1
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
litellm==1.75.0
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
msgpack==1.1.1
multidict==6.6.3
mysql-connector-python==9.4.0
namex==0.1.0
numba==0.61.2
numpy==1.26.4
openai==1.99.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytidb==0.0.11
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.42
sqlmodel==0.0.24
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-hub==0.16.1
termcolor==3.1.0
tf_keras==2.16.0
threadpoolctl==3.6.0
tiktoken==0.10.0
tokenizers==0.21.4
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0